Model validation version 1 (note this code starts in 1950.5 rather than 1950)

+++ person.h
@@ -36,7 +36,7 @@ public:
 	
     double DateOfDeath;								// Variables related to death
 	int Alive;
-	double AgeAtDeath;
+	int AgeAtDeath;
 
 	int HIVStatus;									// Variables related to HIV
 	int MyDateOfHIV;
